requirements.txt
claude-agent-sdk>=0.1.44
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
aiogram>=3.15.0
